// src/components/__generated__/mapUserData.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: mapUserData
// ====================================================

export interface mapUserData {
  __typename: "User";
  id: string;
  firstName: string;
  lastName: string;
  lat: any | null;
  lng: any | null;
  headline: string;
  introduction: string;
}